Break identifier logic tests into their own file
