+++ tests/tests.py
@@ -14,9 +14,7 @@ from caselawclient.Client import (
     MarklogicCommunicationError,
     MarklogicResourceNotFoundError,
 )
-from caselawclient.factories import JudgmentFactory, PressSummaryFactory
 from caselawclient.models.identifiers.neutral_citation import NeutralCitationNumber
-from caselawclient.models.identifiers.press_summary_ncn import PressSummaryRelatedNCNIdentifier
 from caselawclient.models.utilities.aws import S3PrefixString
 from conftest import s3_message_raw, v2_message, v2_message_raw
 from helpers import create_fake_bulk_file, create_fake_tdr_file
@@ -740,33 +738,3 @@ class TestEmailLogic:
 
         updated.assert_not_called()
         new.assert_not_called()
-
-
-class TestDocIdentifiers:
-    def test_select_type_press_summary(self):
-        ingest = MagicMock()
-        ingest.uri = "jam/1/ps/1"
-
-        doc = PressSummaryFactory.build()
-        doc.identifiers = MagicMock()
-        doc.save_identifiers = MagicMock()
-        doc.neutral_citation = "[2013] UKSC 1"
-        ingest.api_client.get_document_by_uri.return_value = doc
-
-        ingester.Ingest.set_document_identifiers(ingest)
-        assert type(doc.identifiers.add.call_args_list[0].args[0]) is PressSummaryRelatedNCNIdentifier
-        doc.save_identifiers.assert_called()
-
-    def test_select_type_judgment(self):
-        ingest = MagicMock()
-        ingest.uri = "jam/1"
-
-        doc = JudgmentFactory.build()
-        doc.identifiers = MagicMock()
-        doc.save_identifiers = MagicMock()
-        doc.neutral_citation = "[2013] UKSC 1"
-        ingest.api_client.get_document_by_uri.return_value = doc
-
-        ingester.Ingest.set_document_identifiers(ingest)
-        assert type(doc.identifiers.add.call_args_list[0].args[0]) is NeutralCitationNumber
-        doc.save_identifiers.assert_called()
